tpm2: Add test harness for marshalling and reconstructing state
Add a test harness that takes the current NvChip and converts it into a
byte stream and then takes the bytes stream and converts it back into the
NvChip memory and compares the result for differences.

To enable the continous testing, one must enable '#define RUN_TESTS' in
NVMem.c.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>